+++ wn/_types.py
@@ -1,5 +1,5 @@
 
-from typing import Union, Mapping, Sequence, Dict, Any
+from typing import Union, Callable, Mapping, Sequence, Dict, Any
 from pathlib import Path
 
 # For functions taking a filesystem path as a str or a pathlib.Path
@@ -10,3 +10,6 @@ RelationMap = Mapping[str, Sequence[str]]
 
 # User-facing metadata representation
 Metadata = Dict[str, Any]
+
+# A function that returns a normalized word form for a given word form
+NormalizeFunction = Callable[[str], str]